rename column_chunk file to column_chunk_data
