+++ src/include/processor/operator/persistent/rel_batch_insert.h
@@ -4,7 +4,6 @@
 #include "processor/operator/partitioner.h"
 #include "processor/operator/persistent/batch_insert.h"
 #include "storage/store/chunked_node_group.h"
-#include "storage/store/column_chunk.h"
 
 namespace kuzu {
 namespace processor {
