+++ src/include/storage/store/chunked_node_group.h
@@ -3,7 +3,7 @@
 #include "common/column_data_format.h"
 #include "common/constants.h"
 #include "common/copy_constructors.h"
-#include "storage/store/column_chunk.h"
+#include "storage/store/column_chunk_data.h"
 
 namespace kuzu {
 namespace storage {
