+++ src/include/storage/store/string_chunk_data.h
@@ -3,8 +3,8 @@
 #include "common/assert.h"
 #include "common/data_chunk/sel_vector.h"
 #include "common/types/types.h"
-#include "storage/store/column_chunk.h"
 #include "storage/store/dictionary_chunk.h"
+#include "storage/store/string_chunk_data.h"
 
 namespace kuzu {
 namespace storage {
