+++ src/processor/operator/persistent/index_builder.cpp
@@ -9,7 +9,7 @@
 #include "common/type_utils.h"
 #include "common/types/ku_string.h"
 #include "storage/index/hash_index_utils.h"
-#include "storage/store/string_column_chunk.h"
+#include "storage/store/string_chunk_data.h"
 
 namespace kuzu {
 namespace processor {
