+++ src/storage/compression/compression.cpp
@@ -19,7 +19,7 @@
 #include "storage/compression/bitpacking_utils.h"
 #include "storage/compression/sign_extend.h"
 #include "storage/storage_utils.h"
-#include "storage/store/column_chunk.h"
+#include "storage/store/column_chunk_data.h"
 #include <ranges>
 
 using namespace kuzu::common;
