+++ src/storage/store/string_column.cpp
@@ -6,7 +6,7 @@
 #include "storage/compression/compression.h"
 #include "storage/store/column.h"
 #include "storage/store/null_column.h"
-#include "storage/store/string_column_chunk.h"
+#include "storage/store/string_chunk_data.h"
 #include "transaction/transaction.h"
 
 using namespace kuzu::catalog;
