+++ src/storage/store/struct_chunk_data.cpp
@@ -1,10 +1,8 @@
-#include "storage/store/struct_column_chunk.h"
+#include "storage/store/struct_chunk_data.h"
 
 #include "common/data_chunk/sel_vector.h"
 #include "common/types/internal_id_t.h"
 #include "common/types/types.h"
-#include "storage/store/column_chunk.h"
-#include <storage/store/string_column_chunk.h>
 
 using namespace kuzu::common;
